Update date input
The three entry fields for the dates (the initial date, the player's birth date and the time change date) have changed. Now, each date field is composed by a normal text field and next to it there is a button that opens the calendar date picker. If you change the date from the calendar, the text field will be updated; if you change the date from the text field, the calendar will be updated.
The date format in the text field must be ISO 8601 (yyyy-mm-dd) otherwise it will not allow you to continue.