Calendar button added and picker hidden
Now, in the three date input ("Data di inizio", "Data di nascita" in the player's details and "Scorri tempo" at the round ending) there is a normal text field which must have the correct date format. Next to it there is a calendar button, when clicked it reveals the calendar date picker.
In addition, I have adjusted the layout by using the already existing classes.