Combat update
At the end of a combat there is a new slide where you can choose which players are dead during the combat: they will be removed from the schedule and from the player list in the next rounds (but not from the player array).
In addition, the report at the end of the game is more clear about the combat time and the dead player list has been fixed.